misc/Gotcha/setup-gotcha/app/requirements.txt
Pillow
flask
gunicorn